put native post in a separate route
